requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
python-dotenv
plaid-python
requests
Flask-Bootstrap
chart.js
gunicorn